requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1

pytorch-lightning==2.4.0
tabulate>=0.9.0
tqdm>=4.66.5
pillow>=10.4.0
notebook>=7.2.2
jupyterlab>=4.2.5
matplotlib>=3.9.2
ipywidgets>=8.1.2
plotly==5.17.0
einops==0.8.0
numpy==2.2.1
pandas==2.2.3
loguru==0.7.3
python-dotenv==1.0.1
pygraphviz==1.14
huggingface-hub==0.27.1

opencv-python
numpy==1.23
open3d
object2urdf
trimesh==3.10
pybullet
mediapy
plyfile
setuptools==49.4
hydra-core>=1.3.2
pynput
